Remove requirement for location to be zip code
Nominatim will take a huge variety of descriptions for the location and
is generally pretty good at getting the right place. This commit simply
changes the logic so any location can be provided by the user, checking
that is is a real location AND the location the user intended should be
included.